requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
alembic==1.14.0
pydantic==2.10.3
pydantic-settings==2.6.1
python-multipart==0.0.19
python-jose[cryptography]==3.3.0
bcrypt==4.1.2
pillow==11.0.0
torch==2.5.1
torchvision==0.20.1
numpy==2.2.0
aiosqlite==0.20.0
python-dotenv==1.0.1
